- Prevent non-existing file errors

+++ APPropertyWrappers/Core/Simple/FilePreserved.swift
@@ -84,8 +84,10 @@ open class FilePreserved {
     
     /// Resets value to its default by erasing related file.
     public func reset() {
+        storage = nil
+        guard FileManager.default.fileExists(atPath: url.path) else { return }
+        
         do {
-            storage = nil
             try FileManager.default.removeItem(at: url)
         } catch {
             RoutableLogger.logError("Unable to reset preserved data", error: error, data: ["url": url])
